+++ package.json
@@ -12,12 +12,11 @@
     "@sanity/block-content-to-react": "^1.3.12",
     "@sanity/client": "^0.132.10",
     "@sanity/image-url": "^0.132.8",
-    "@zeit/next-sass": "^1.0.1",
     "cross-env": "^6.0.3",
     "es6-promise": "^4.2.8",
     "exenv": "^1.2.2",
     "isomorphic-fetch": "^2.2.1",
-    "next": "^9.0.0",
+    "next": "^9.2",
     "next-cookies": "^2.0.1",
     "node-sass": "^4.12.0",
     "react": "^16.9.0",
