+++ themes/victorian/victorian.module.css
@@ -1,4 +1,8 @@
-body {
+.victorian {
 	--color__branding-primary: #fff;
 	background-color: green;
+}
+
+.victorian p {
+	color: blue;
 }
